feat: add main method fix: lapTime must be an integer

+++ src/net/jeremycastel/anotherfano/fanos/Bike.java
@@ -1,7 +1,7 @@
 package net.jeremycastel.anotherfano.fanos;
 
 public class Bike extends Fano {
-    public Bike(int number, double lapTime) throws Exception {
+    public Bike(int number, int lapTime) throws Exception {
         super(number, lapTime);
     }
 }
