+++ src/net/jeremycastel/anotherfano/fanos/Car.java
@@ -1,7 +1,7 @@
 package net.jeremycastel.anotherfano.fanos;
 
 public class Car extends Fano {
-    public Car(int number, double lapTime) throws Exception {
+    public Car(int number, int lapTime) throws Exception {
         super(number, lapTime);
     }
 }
