+++ src/net/jeremycastel/anotherfano/fanos/Kart.java
@@ -1,7 +1,7 @@
 package net.jeremycastel.anotherfano.fanos;
 
 public class Kart extends Fano {
-    public Kart(int number, double lapTime) throws Exception {
+    public Kart(int number, int lapTime) throws Exception {
         super(number, lapTime);
     }
 }
